+++ megamek/src/megamek/common/pathfinder/AeroSpacePathFinder.java
@@ -94,7 +94,8 @@ public class AeroSpacePathFinder extends NewtonianAerospacePathFinder {
         
         // terminator conditions:
         // we've visited this hex already and the path we are considering is longer than the previous path that visited this hex
-        if (visitedCoords.containsKey(pathDestination) && visitedCoords.get(pathDestination).intValue() < path.getMpUsed()) {
+        if (visitedCoords.containsKey(pathDestination)
+                && (visitedCoords.get(pathDestination) < path.getMpUsed())) {
             return true;
         }
